[LC-14] added hook form to signup

+++ src/modules/auth/component/SignUp/SignUp.tsx
@@ -1,58 +1,66 @@
 import {
   Grid,
+  Typography,
 } from "@mui/material";
-import {
-  ChangeEvent, FormEvent, useEffect, useState,
-} from "react";
+import { useEffect } from "react";
 import { v4 as uuidv4 } from "uuid";
+import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
 import { PrimaryButton } from "../../../../components/UI/PrimaryButton";
-import { TextField } from "../../../../components/UI/TextField";
 import { Heading } from "../../../../components/UI/Heading";
-import { profileData, token } from "../../../../data-mock/data";
+import { token } from "../../../../data-mock/data";
 import { Profile } from "../../../../types/types";
 import { useAppDispatch } from "../../../../store/hooks";
 import { login } from "../../../../store/features/tokenSlice";
 import { UserRepository, userRepository } from "../../../../data-mock/userMock";
+import { UserForm } from "../../../../components/ConnectForm/UserForm";
 
 export function SignUp() {
-  const [formData, setFormData] = useState<Omit<Profile, "id" | "isLoggedIn">>(
-    profileData,
-  );
+  const methods = useForm<Omit<Profile, "id">>({
+    defaultValues: {
+      name: "",
+      email: "",
+      password: "",
+      confirmPassword: "",
+    },
+  });
+
+  const {
+    handleSubmit, setError, reset, formState: { errors },
+  } = methods;
   const users: UserRepository = userRepository;
   const dispatch = useAppDispatch();
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData({
-      ...formData,
-      [name]: value,
-    });
-  };
 
   useEffect(() => {
     // eslint-disable-next-line no-console
     console.log("All Submitted Data:", users);
   }, [users]);
 
-  const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const newFormData: Profile = {
-      id: uuidv4(),
-      ...formData,
-    };
-
-    await users.addUser(newFormData);
-    dispatch(login(token));
-    localStorage.setItem("user", JSON.stringify({ email: newFormData.email }));
-    localStorage.setItem("token", token);
-    setFormData(
-      {
-        name: "",
-        email: "",
-        password: "",
-        confirmPassword: "",
-      },
+  const onSubmit: SubmitHandler<Omit<Profile, "id">> = async (data) => {
+    const userExists = await users.getUserByEmail(
+      data.email,
     );
+
+    if (userExists) {
+      setError("email", { type: "manual", message: "User with such email is already registered" });
+    } else {
+      const newFormData: Profile = {
+        id: uuidv4(),
+        ...data,
+      };
+
+      await users.addUser(newFormData);
+      dispatch(login(token));
+      localStorage.setItem("user", JSON.stringify({ email: newFormData.email }));
+      localStorage.setItem("token", token);
+      reset(
+        {
+          name: "",
+          email: "",
+          password: "",
+          confirmPassword: "",
+        },
+      );
+    }
   };
 
   return (
@@ -73,62 +81,59 @@ export function SignUp() {
       >
         <Grid container spacing={2}>
           <Grid item sm={12}>
+            <FormProvider {...methods}>
 
-            <form onSubmit={handleSubmit}>
-              <Grid container spacing={6}>
-                <Grid item xs={12}>
-                  <TextField
-                    label="Email"
-                    placeholder="email@gmail.com"
-                    type="email"
-                    name="email"
-                    value={formData.email}
-                    onChange={handleChange}
-                  />
+              <form onSubmit={handleSubmit(onSubmit)}>
+                <Grid container spacing={6}>
+                  <Grid item xs={12}>
+                    <UserForm
+                      name="email"
+                      label="Email"
+                      placeholder="email@gmail.com"
+                      type="email"
+                    />
+                  </Grid>
+                  <Grid item xs={12}>
+                    <UserForm
+                      name="name"
+                      label="Seller name"
+                      placeholder="Seller name"
+                      type="text"
+                    />
+                  </Grid>
+                  <Grid item xs={12}>
+                    <UserForm
+                      name="password"
+                      label="Password"
+                      placeholder="********"
+                      type="password"
+                    />
+                  </Grid>
+                  <Grid item xs={12}>
+                    <UserForm
+                      name="confirmPassword"
+                      label="Confirm password"
+                      placeholder="********"
+                      type="password"
+                    />
+                  </Grid>
+                  {errors.email && (
+                    <Typography color="red" paddingLeft="48px">{errors.email.message}</Typography>
+                  )}
+                  {errors.password && (
+                    <Typography color="red" paddingLeft="48px">{errors.password.message}</Typography>
+                  )}
+                  {errors.confirmPassword && (
+                    <Typography color="red" paddingLeft="48px">{errors.confirmPassword.message}</Typography>
+                  )}
+                  <Grid item xs={12}>
+                    <PrimaryButton>
+                      Submit
+                    </PrimaryButton>
+                  </Grid>
                 </Grid>
-                <Grid item xs={12}>
-                  <TextField
-                    label="Seller name"
-                    placeholder="Seller name"
-                    type="text"
-                    size="small"
-                    fullWidth
-                    name="name"
-                    value={formData.name}
-                    onChange={handleChange}
-                  />
-                </Grid>
-                <Grid item xs={12}>
-                  <TextField
-                    label="Password"
-                    placeholder="********"
-                    type="password"
-                    size="small"
-                    fullWidth
-                    name="password"
-                    value={formData.password}
-                    onChange={handleChange}
-                  />
-                </Grid>
-                <Grid item xs={12}>
-                  <TextField
-                    label="Confirm password"
-                    placeholder="********"
-                    type="password"
-                    size="small"
-                    fullWidth
-                    name="confirmPassword"
-                    value={formData.confirmPassword}
-                    onChange={handleChange}
-                  />
-                </Grid>
-                <Grid item xs={12}>
-                  <PrimaryButton>
-                    Submit
-                  </PrimaryButton>
-                </Grid>
-              </Grid>
-            </form>
+              </form>
+            </FormProvider>
           </Grid>
         </Grid>
       </Grid>
